Apply a few physical fixes + 1 minor optimization
The optimization is from a rather surprising
profiling result (Visual Studio), expected to
save about 1-2% of CPU time.

The physical fix is that I had forgotten to take the
mass of the receiving particle when calculating
the aceleration due to gravity. I've added a step
of division by mass (a = F/m).