Refactor; fix integration; tweak.
Refactoring:
1. Mostly renaming, const & stuff.
2. Hide internal state of Halton.
3. Add comments.

Integration:
1. Suppress potential overflow in counting inaccuracies, etc.
Due to O(N^2) time complexity (where N is the number
of particles), and because "int" is usually 32-bit wide,
The counters could overflow with sqrt(2^32)
or more particles (about that size), or about 2^16
particles. So, relatively few particles could overflow that
number. But I'm not using absolute counts at all.
So I used booleans instead.
2. Proper handling of neutral decision:
Because of a programming error, the neutral decision
(time step not recommended to change)
was being handled as though it were a pessimistic
decision (time step advised to shrink). This
also slowed down the program because a
pessimistic decision would cause the computation
to be retried.

Tweak:
One notable thing: I set a lower bound for
the sizes and masses of particles, so there
are fewer particles moving too quickly.
When particles move too quickly,
accuracy protection kicks in, which slows
down the time step and retries the affected
interaction.